+++ NEWS.md
@@ -1,7 +1,7 @@
 # odytools 0.7.5.9000
 
 -   Added **ody_gt2image** to convert a GT table to an image.
--   When accessing a project (both plain and REDCap-based), the existence and update status of the renv.lock file are now checked. If it is not present or outdated, the user is prompted to create or update it.
+-   When accessing a project (both plain and REDCap-based), the existence and update status of the renv.lock file and a git repo are now checked.
 -   When a project is started (both plain and REDCap-based), a template .gitignore file is now added.
 -   In the REDCap Data Viewer, the *Event* dropdown list now also shows the Arm (if any) the event belongs to.
 -   **ody_rc_import** now includes the attributes: